The final Upadate

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: mbari <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/17 17:20:00 by mbari             #+#    #+#             */
-/*   Updated: 2019/10/18 17:57:21 by mbari            ###   ########.fr       */
+/*   Updated: 2019/10/30 12:06:52 by mbari            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *str)
+size_t		ft_strlen(const char *str)
 {
 	size_t	i;
